🌱 Extract PodDrillDown types, cache, and helpers
Extract ~95 lines from PodDrillDown.tsx (2,374→2,279 lines) into
a new `pod-drilldown/` subdirectory with barrel exports:

- types.ts: PodDrillDownProps, TabType, RelatedResource, CachedData,
  UNHEALTHY_STATUSES
- helpers.ts: getIssueSeverity (issue severity classification)
- podCache.ts: module-level pod data cache with auto-cleanup
- index.ts: barrel re-exports
